MINIFICPP-517: Updates to PR based on comments
