+++ extensions/sensors/CMakeLists.txt
@@ -17,13 +17,8 @@
 # under the License.
 #
 
-cmake_minimum_required(VERSION 2.6)
+include(${CMAKE_SOURCE_DIR}/extensions/ExtensionHeader.txt)
 
-
-set(CMAKE_EXE_LINKER_FLAGS "-Wl,--export-all-symbols")
-set(CMAKE_SHARED_LINKER_FLAGS "-Wl,--export-symbols")
-
-include_directories(../../libminifi/include  ../../libminifi/include/core/yaml  ../../libminifi/include/core  ../../thirdparty/spdlog-20170710/include ../../thirdparty/concurrentqueue ../../thirdparty/yaml-cpp-yaml-cpp-0.5.3/include ${CIVET_THIRDPARTY_ROOT}/include ../../thirdparty/) 
 include_directories(../../thirdparty/RTIMULib/RTIMULib)
 file(GLOB SOURCES  "*.cpp")
