+++ extensions/sensors/GetEnvironmentalSensors.h
@@ -32,6 +32,7 @@
 #include "core/Resource.h"
 #include "utils/Id.h"
 #include "RTIMULib.h"
+#include "SensorBase.h"
 #include "RTMath.h"
 
 namespace org {
@@ -41,7 +42,7 @@ namespace minifi {
 namespace processors {
 
 // EnvironmentalSensors Class
-class GetEnvironmentalSensors : public core::Processor {
+class GetEnvironmentalSensors : public SensorBase {
  public:
 
   // Constructor
@@ -49,8 +50,7 @@ class GetEnvironmentalSensors : public core::Processor {
    * Create a new processor
    */
   GetEnvironmentalSensors(std::string name, uuid_t uuid = NULL)
-      : Processor(name, uuid),
-        imu(nullptr),
+      : SensorBase(name, uuid),
         humidity_sensor_(nullptr),
         pressure_sensor_(nullptr),
         logger_(logging::LoggerFactory<GetEnvironmentalSensors>::getLogger()) {
@@ -66,24 +66,9 @@ class GetEnvironmentalSensors : public core::Processor {
   virtual void initialize() override;
   virtual void onSchedule(const std::shared_ptr<core::ProcessContext> &context, const std::shared_ptr<core::ProcessSessionFactory> &sessionFactory) override;
 
-  class WriteCallback : public OutputStreamCallback {
-     public:
-      WriteCallback(std::string data)
-          : _data(const_cast<char*>(data.data())),
-            _dataSize(data.size()) {
-      }
-      char *_data;
-      uint64_t _dataSize;
-      int64_t process(std::shared_ptr<io::BaseStream> stream) {
-        int64_t ret = 0;
-        if (_data && _dataSize > 0)
-          ret = stream->write(reinterpret_cast<uint8_t*>(_data), _dataSize);
-        return ret;
-      }
-    };
+ protected:
+  virtual void notifyStop();
  private:
-  RTIMUSettings settings;
-  RTIMU *imu;
   RTHumidity *humidity_sensor_;
   RTPressure *pressure_sensor_;
   std::shared_ptr<logging::Logger> logger_;
